requirements.txt
fastapi==0.99.0
uvicorn[standard]==0.22.0
playwright==1.39.2
httpx==0.24.0
pydantic==2.5.0
google-generativeai==0.7.0
pdfplumber==0.9.0
pillow==10.0.0
pytesseract==0.3.10
pandas==2.2.2
openpyxl==3.1.2
